requirements.txt
fpdf
python-docx
# pypandoc (optional, for more formats)
